+++ spec/Glitch/Runtime/AssignmentExceptionSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Glitch\Interpreter;
+namespace spec\Glitch\Runtime;
 
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
